Add files via upload
Zebras_Sikora was an assignment in my Intro to Computer Science 2 class where we were given the file ZebraBotswana.txt file, which has information on 7 different zebra's migration patterns.  This information was given to us using a Unix timestamp, longitude, and latitude.  Our assignment was to create a program in python that uses the pylab package to plot all the zebras' longitude and latitude positions.  We were also tasked to calculate the average distance traveled per day and display it on a bar graph.

wheatseeds_Sikora was an assign in my Intro to Computer Science 2 class where we took the seedsData.txt file, which contains data on randomly selected kernels of wheat belonging to three different varieties of wheat: Kama, Rosa, and Canadian.  Each line has 8 values, the first 7 are attributes about the kernel (in order: area, perimeter, compactness, length of kernel, width of kernel, asymmetry coefficient, and length of kernel groove), and the last value is the type of kernel (1=Kama, 2=Rosa, 3=Canadian).  Our goal for the program was to use k-means clustering to correctly partition the set of kernels into the three known varieties.  Then we created functions to evaluate our clusters, which were the cohesion (the reciprocal of the sum of the distances from each point in the cluster to the centroid of the cluster), the separation (the distance between each cluster's centroid), and the integrity (the proportion of the cluster comprised by the majority type in the cluster)

Cloudera Documentation_Sikora was part of a take home exam in my Business Intelligence and Big Data class where I worked through a tutorial in Hadoop using a Cloudera virtual machine (the file Exam Part 2 - Hadoop is the description of the test).  The documentation file shows my step by step process of going through the tutorials, some of the problems I encountered as well as how I solved them.  This shows that I have been introduced to the Hadoop environment and have experience not only with the different Hadoop applications, but also troubleshooting and problem solving.  There is an ending reflection that shows my initial thoughts of working in Hadoop, and at the very bottom in the red text are my instructor's comments on my test and the grade I received on it.